+++ src/playground/editor.jsx
@@ -20,8 +20,8 @@ import ReactDOM from 'react-dom';
 import React from 'react';
 import {setAppElement} from 'react-modal';
 
-import Interface from './render-interface.jsx';
+import Playground from './playground.jsx';
 import appTarget from './app-target';
 
 setAppElement(appTarget);
-ReactDOM.render(<Interface />, appTarget);
+ReactDOM.render(<Playground />, appTarget);
