+++ src/playground/player.jsx
@@ -20,10 +20,10 @@ import ReactDOM from 'react-dom';
 import React from 'react';
 import {setAppElement} from 'react-modal';
 
-import Interface from './render-interface.jsx';
+import Playground from './playground.jsx';
 import appTarget from './app-target';
 
 setAppElement(appTarget);
-ReactDOM.render(<Interface
+ReactDOM.render(<Playground
     isPlayerOnly
 />, appTarget);
